rename: #103 rename asset icons
