+++ src/components/buttons/imageButton/ImageButton.tsx
@@ -5,9 +5,9 @@ import { Shadow } from 'react-native-shadow-2'
 
 import DesignIcon from '@/assets/icons/design.svg'
 import FoodIcon from '@/assets/icons/food.svg'
-import HomeLivingIcon from '@/assets/icons/homeLiving.svg'
+import HomeLivingIcon from '@/assets/icons/home-living.svg'
 import LifeIcon from '@/assets/icons/life.svg'
-import SelfImprovementIcon from '@/assets/icons/selfImprovement.svg'
+import SelfImprovementIcon from '@/assets/icons/self-improvement.svg'
 import ShoppingIcon from '@/assets/icons/shopping.svg'
 import SportIcon from '@/assets/icons/sport.svg'
 import TechIcon from '@/assets/icons/tech.svg'
