+++ src/components/cards/archivingCard/ArchivingCard.tsx
@@ -11,8 +11,8 @@ import { deleteArchiving, patchPinArchiving, patchScrapArchiving } from '@/apis/
 import { defaultIcons, defaultImages } from '@/assets'
 import PhotoIcon from '@/assets/icons/photo.svg'
 import ScrapIcon from '@/assets/icons/scrap.svg'
-import ScrapFillIcon from '@/assets/icons/scrap_fill.svg'
-import ScrapSmallIcon from '@/assets/icons/scrap_small.svg'
+import ScrapFillIcon from '@/assets/icons/scrap-fill.svg'
+import ScrapSmallIcon from '@/assets/icons/scrap-small.svg'
 import TwoButtonDialog from '@/components/dialogs/twoButtonDialog/TwoButtonDialog'
 import Popup from '@/components/popup/Popup'
 import { ArchivingInfo } from '@/models/Archiving'
