+++ src/components/cards/popularArchivingCard/PopularArchivingCard.tsx
@@ -9,8 +9,8 @@ import { patchScrapArchiving } from '@/apis/archiving/Archiving'
 import { defaultIcons, defaultImages } from '@/assets'
 import PhotoIcon from '@/assets/icons/photo.svg'
 import ScrapIcon from '@/assets/icons/scrap.svg'
-import ScrapFillIcon from '@/assets/icons/scrap_fill.svg'
-import ScrapSmallIcon from '@/assets/icons/scrap_small.svg'
+import ScrapFillIcon from '@/assets/icons/scrap-fill.svg'
+import ScrapSmallIcon from '@/assets/icons/scrap-small.svg'
 import { ArchivingInfo } from '@/models/Archiving'
 import { MainNavigationProp } from '@/navigations/MainNavigator'
 import { CommunityCategoryState } from '@/state/CategoryState'
