+++ src/components/modal/archivingModal/createArchivingModal/CreateArchivingModal.tsx
@@ -17,7 +17,7 @@ import { useRecoilState, useRecoilValue } from 'recoil'
 import { postArchiving } from '@/apis/archiving/Archiving'
 import { defaultImages } from '@/assets'
 import CameraIcon from '@/assets/icons/camera.svg'
-import XMark from '@/assets/icons/x_mark.svg'
+import XMark from '@/assets/icons/x-mark.svg'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import { DropDown } from '@/components/dropDown/DropDown'
 import i18n from '@/locales'
