+++ src/components/modal/archivingModal/editArchivingModal/EditArchivingModal.tsx
@@ -17,7 +17,7 @@ import { useRecoilState, useRecoilValue } from 'recoil'
 import { getArchivingData, patchArchiving } from '@/apis/archiving/Archiving'
 import { defaultImages } from '@/assets'
 import CameraIcon from '@/assets/icons/camera.svg'
-import XMark from '@/assets/icons/x_mark.svg'
+import XMark from '@/assets/icons/x-mark.svg'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import { InformationErrorDialog } from '@/components/dialogs/errorDialog/InformationErrorDialog/InformationErrorDialog'
 import { DropDown } from '@/components/dropDown/DropDown'
