+++ src/components/modal/selectArchivingModal/SelectArchivingModal.tsx
@@ -6,9 +6,9 @@ import { useQuery } from 'react-query'
 import { useRecoilState, useSetRecoilState } from 'recoil'
 
 import { getArchivingList } from '@/apis/archiving/ArchivingList'
-import CheckCircle from '@/assets/icons/check_circle.svg'
-import LeftProvider from '@/assets/icons/left_provider.svg'
-import XMark from '@/assets/icons/x_mark.svg'
+import CheckCircle from '@/assets/icons/check-circle.svg'
+import LeftProvider from '@/assets/icons/left-provider.svg'
+import XMark from '@/assets/icons/x-mark.svg'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import DefaultScrollContainer from '@/components/containers/defaultScrollContainer/DefaultScrollContainer'
 import { Divider } from '@/components/divider/Divider'
