+++ src/screens/contentDetail/components/imageDetail/components/ImageHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import XMark from '@/assets/icons/x_mark.svg'
+import XMark from '@/assets/icons/x-mark.svg'
 import { colors } from '@/styles/colors'
 
 import { CloseButton, Container, Header, Title } from './ImageHeader.style'
