+++ src/screens/contentList/ContentList.tsx
@@ -12,8 +12,8 @@ import { deleteArchiving, patchScrapArchiving } from '@/apis/archiving/Archiving
 import { getContentByArchiving } from '@/apis/archiving/ArchivingList'
 import { postBlock } from '@/apis/block/Block'
 import { defaultImages } from '@/assets'
+import ScrapFillIcon from '@/assets/icons/scrap-fill.svg'
 import ScrapIcon from '@/assets/icons/scrap.svg'
-import ScrapFillIcon from '@/assets/icons/scrap_fill.svg'
 import ContentCard from '@/components/cards/contentCard/ContentCard'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
 import DefaultDialog from '@/components/dialogs/defaultDialog/DefaultDialog'
