+++ src/screens/mypage/Mypage.tsx
@@ -10,7 +10,7 @@ import { useMutation, useQuery, useQueryClient } from 'react-query'
 import { logout } from '@/apis/auth/Auth'
 import { getUser } from '@/apis/user/User'
 import { defaultImages } from '@/assets'
-import LeftArrowIcon from '@/assets/icons/left_arrow.svg'
+import LeftArrowIcon from '@/assets/icons/left-arrow.svg'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
 import { InformationErrorDialog } from '@/components/dialogs/errorDialog/InformationErrorDialog/InformationErrorDialog'
 import { Loading } from '@/components/loading/Loading'
